Fix workspacemenucontainer test
