+++ src/subapps/studioLegacy/containers/__tests__/WorkSpaceMenuContainer.spec.tsx
@@ -1,7 +1,5 @@
 import '@testing-library/jest-dom';
 import { vi } from 'vitest';
-import { rest } from 'msw';
-import { act } from 'react-dom/test-utils';
 import { NexusProvider } from '@bbp/react-nexus';
 import { createBrowserHistory } from 'history';
 import { createNexusClient } from '@bbp/nexus-sdk';
@@ -14,7 +12,6 @@ import { configureStore } from '../../../../store';
 import {
   render,
   fireEvent,
-  waitFor,
   screen,
   server,
 } from '../../../../utils/testUtil';
@@ -22,6 +19,10 @@ import StudioReactContext, {
   StudioContextType,
 } from '../../contexts/StudioContext';
 import { deltaPath } from '__mocks__/handlers/handlers';
+import { aclHandler, dashboardHandler, tableHandler, workspaceHandler } from './WorkSpaceMenuContainerHandlers';
+import userEvent from '@testing-library/user-event';
+import { UserEvent } from '@testing-library/user-event/dist/types/setup/setup';
+import { RenderResult } from '@testing-library/react';
 
 describe('workSpaceMenu', () => {
   const history = createBrowserHistory({ basename: '/' });
@@ -80,376 +81,28 @@ describe('workSpaceMenu', () => {
     _updatedBy:
       'https://dev.nise.bbp.epfl.ch/nexus/v1/realms/local/users/localuser',
   } as unknown) as StudioResource;
+
   const queryClient = new QueryClient();
+
+  let container: HTMLElement;
+  let user: UserEvent;
+  let component: RenderResult;
+
   // establish API mocking before all tests
-  beforeAll(() => server.listen());
-  // reset any request handlers that are declared as a part of our tests
-  // (i.e. for testing one-time error scenarios)
-  afterEach(() => server.resetHandlers());
-  // clean up once the tests are done
-  afterAll(() => server.close());
-  server.use(
-    rest.get(deltaPath('/resources/org/project/_/w1'), (req, res, ctx) => {
-      const mockResponse = {
-        '@context': [
-          'https://bluebrain.github.io/nexus/contexts/metadata.json',
-          'https://bluebrainnexus.io/studio/context',
-        ],
-        '@id': 'w1',
-        '@type': 'StudioWorkSpace',
-        description: `A test work space`,
-        label: `w1`,
-        dashboards: [
-          {
-            dashboard: 'd1',
-          },
-        ],
-      };
-      return res(
-        // Respond with a 200 status code
-        ctx.status(200),
-        ctx.json(mockResponse)
-      );
-    })
-  );
+  beforeAll(() => {
+    server.listen();
+  });
 
-  server.use(
-    rest.get(deltaPath('/resources/org/project/_/d1'), (req, res, ctx) => {
-      const mockResponse = {
-        '@context': [
-          'https://bluebrain.github.io/nexus/contexts/metadata.json',
-          'https://bluebrainnexus.io/studio/context',
-        ],
-        '@id': 'd1',
-        '@type': 'StudioDashboard',
-        description: `A test dashboard`,
-        label: `d1`,
-        dataTable: {
-          '@id': 'dataTable1',
-        },
-      };
-      return res(
-        // Respond with a 200 status code
-        ctx.status(200),
-        ctx.json(mockResponse)
-      );
-    })
-  );
-  server.use(
-    rest.get(
-      deltaPath('/resources/org/project/_/dataTable1'),
-      (req, res, ctx) => {
-        const mockResponse = {
-          '@context': [
-            'https://bluebrain.github.io/nexus/contexts/metadata.json',
-            'https://bluebrainnexus.io/studio/context',
-          ],
-          '@id': 'dataTable1',
-          '@type': 'FusionTable',
-          description: `A test dataTable`,
-          label: `dataTable1`,
-          configuration: [
-            {
-              '@type': 'text',
-              enableFilter: false,
-              enableSearch: false,
-              enableSort: false,
-              format: '',
-              name: 'o',
-            },
-            {
-              '@type': 'text',
-              enableFilter: false,
-              enableSearch: false,
-              enableSort: false,
-              format: '',
-              name: 'p',
-            },
-          ],
-          dataQuery:
-            'SELECT DISTINCT ?self ?p ?o \nWHERE {\n     ?self ?p ?o .\n}\nLIMIT 10',
-          enableDownload: true,
-          enableInteractiveRows: true,
-          enableSave: true,
-          enableSearch: true,
-          name: 'Example',
-          resultsPerPage: 5,
-          view:
-            'https://bluebrain.github.io/nexus/vocabulary/defaultSparqlIndex',
-        };
-        return res(
-          // Respond with a 200 status code
-          ctx.status(200),
-          ctx.json(mockResponse)
-        );
-      }
-    )
-  );
-  it('renders with a single workspace and dashboard ', async () => {
-    await act(async () => {
-      const { container } = await render(
-        <Provider store={store}>
-          <ConnectedRouter history={history}>
-            <NexusProvider nexusClient={nexus}>
-              <QueryClientProvider client={queryClient}>
-                <StudioReactContext.Provider value={contextValue}>
-                  <WorkSpaceMenu
-                    workspaceIds={['w1']}
-                    studioResource={resource}
-                    onListUpdate={vi.fn}
-                  ></WorkSpaceMenu>
-                </StudioReactContext.Provider>
-              </QueryClientProvider>
-            </NexusProvider>
-          </ConnectedRouter>
-        </Provider>
-      );
-      await waitFor(async () => {
-        const dashboardText = await screen.getAllByText(/Example/);
-        expect(dashboardText.length).toBe(1);
-        const text = await screen.getAllByText(
-          'https://bluebrain.github.io/nexus/vocabulary/apiMappings'
-        );
-        expect(text.length).toBe(9);
-      });
-      expect(container).toMatchSnapshot();
-    });
-  });
-  it('shows edit buttons for a user with  edit access', async () => {
-    await act(async () => {
-      await render(
-        <Provider store={store}>
-          <ConnectedRouter history={history}>
-            <NexusProvider nexusClient={nexus}>
-              <QueryClientProvider client={queryClient}>
-                <StudioReactContext.Provider value={contextValue}>
-                  <WorkSpaceMenu
-                    workspaceIds={['w1']}
-                    studioResource={resource}
-                    onListUpdate={vi.fn}
-                  />
-                </StudioReactContext.Provider>
-              </QueryClientProvider>
-            </NexusProvider>
-          </ConnectedRouter>
-        </Provider>
-      );
-      await waitFor(async () => {
-        const button1 = await screen.getByText('Workspace');
-        expect(button1).toBeVisible();
-        const button2 = await screen.getByText('Dashboard');
-        expect(button2).toBeVisible();
-      });
-    });
-  });
-  it('hides edit buttons for a user with out edit access', async () => {
-    rest.get(deltaPath('/acls/org/project'), (req, res, ctx) => {
-      const mockResponse = {
-        '@context': [
-          'https://bluebrain.github.io/nexus/contexts/metadata.json',
-          'https://bluebrain.github.io/nexus/contexts/search.json',
-          'https://bluebrain.github.io/nexus/contexts/acls.json',
-        ],
-        _total: 1,
-        _results: [
-          {
-            '@id': deltaPath('/v1/acls/org1'),
-            '@type': 'AccessControlList',
-            acl: [
-              {
-                identity: {
-                  '@id': deltaPath('/v1/realms/myrealm/groups/a-group'),
-                  '@type': 'Group',
-                  group: 'a-group',
-                  realm: 'myrealm',
-                },
-                permissions: ['projects/read'],
-              },
-              {
-                identity: {
-                  '@id': deltaPath('/v1/realms/realm/groups/some-group'),
-                  '@type': 'Group',
-                  group: 'some-group',
-                  realm: 'realm',
-                },
-                permissions: ['projects/read', 'projects/write'],
-              },
-              {
-                identity: {
-                  '@id': deltaPath('/v1/realms/local/users/localuser'),
-                  '@type': 'User',
-                  realm: 'local',
-                  subject: 'localuser',
-                },
-                permissions: ['resources/read'], // No write permission.
-              },
-            ],
-            _constrainedBy:
-              'https://bluebrain.github.io/nexus/schemas/acls.json',
-            _createdAt: '2021-05-11T11:03:06.071Z',
-            _createdBy: deltaPath('/v1/anonymous'),
-            _deprecated: false,
-            _path: '/org/project',
-            _rev: 1,
-            _self: deltaPath('/v1/acls/org1'),
-            _updatedAt: '2021-05-11T11:03:06.071Z',
-            _updatedBy: deltaPath('/v1/anonymous'),
-          },
-        ],
-      };
-      return res(
-        // Respond with a 200 status code
-        ctx.status(200),
-        ctx.json(mockResponse)
-      );
-    }),
-      await act(async () => {
-        const { container } = await render(
-          <Provider store={store}>
-            <ConnectedRouter history={history}>
-              <NexusProvider nexusClient={nexus}>
-                <QueryClientProvider client={queryClient}>
-                  <StudioReactContext.Provider value={contextValue}>
-                    <WorkSpaceMenu
-                      workspaceIds={['w1']}
-                      studioResource={resource}
-                      onListUpdate={vi.fn}
-                    ></WorkSpaceMenu>
-                  </StudioReactContext.Provider>
-                </QueryClientProvider>
-              </NexusProvider>
-            </ConnectedRouter>
-          </Provider>
-        );
-        await waitFor(async () => {
-          // wait for the workspace to load.
-          await screen.getAllByText('No dashboards available');
-          const buttons = await screen.findAllByRole('button');
-          expect(buttons).toHaveLength(2);
-        });
-      });
-  });
-  it('it displays workspace action on clicking  workspace action button', async () => {
-    await act(async () => {
-      const { container } = await render(
-        <Provider store={store}>
-          <ConnectedRouter history={history}>
-            <NexusProvider nexusClient={nexus}>
-              <QueryClientProvider client={queryClient}>
-                <StudioReactContext.Provider value={contextValue}>
-                  <WorkSpaceMenu
-                    workspaceIds={['w1']}
-                    studioResource={resource}
-                    onListUpdate={vi.fn}
-                  ></WorkSpaceMenu>
-                </StudioReactContext.Provider>
-              </QueryClientProvider>
-            </NexusProvider>
-          </ConnectedRouter>
-        </Provider>
-      );
-      await waitFor(async () => {
-        const button1 = await screen.findByText('Workspace');
-        expect(button1).toBeVisible();
-        const button2 = await screen.findByText('Dashboard');
-        expect(button2).toBeVisible();
-        const x = await fireEvent.click(button1);
-      });
-      await waitFor(async () => {
-        expect(screen.getByText('Edit')).toBeInTheDocument();
-        expect(screen.getByText('Add')).toBeInTheDocument();
-        expect(screen.getByText('Remove')).toBeInTheDocument();
-      });
-    });
-  });
-  it('it displays workspace add form on clicking add workspace button', async () => {
-    await act(async () => {
-      const { container } = await render(
-        <Provider store={store}>
-          <ConnectedRouter history={history}>
-            <NexusProvider nexusClient={nexus}>
-              <QueryClientProvider client={queryClient}>
-                <StudioReactContext.Provider value={contextValue}>
-                  <WorkSpaceMenu
-                    workspaceIds={['w1']}
-                    studioResource={resource}
-                    onListUpdate={vi.fn}
-                  ></WorkSpaceMenu>
-                </StudioReactContext.Provider>
-              </QueryClientProvider>
-            </NexusProvider>
-          </ConnectedRouter>
-        </Provider>
-      );
-    });
+  beforeEach(async () => {
+    server.use(workspaceHandler);
+    server.use(dashboardHandler);
+    server.use(tableHandler);
 
-    await waitFor(async () => {
-      await screen.getAllByText('Workspace');
-    });
-    await act(async () => {
-      const workSpaceAction = (await screen.getAllByText(
-        'Workspace'
-      )[0]) as HTMLButtonElement;
-      fireEvent.click(workSpaceAction);
-    });
-    await waitFor(async () => {
-      await screen.findByText('Add');
-    });
-    const editButton = await screen.findByText('Add');
-    // expect(editButton).toBeVisible();
-    await act(async () => {
-      await fireEvent.click(editButton);
-    });
-    await waitFor(async () => {
-      const editForm = await screen.findAllByText('Create a new Workspace');
-      expect(editForm).toHaveLength(1);
-    });
-  });
-  it('it displays workspace remove confirmation  on clicking remove workspace button', async () => {
-    await act(async () => {
-      const { container } = await render(
-        <Provider store={store}>
-          <ConnectedRouter history={history}>
-            <NexusProvider nexusClient={nexus}>
-              <QueryClientProvider client={queryClient}>
-                <StudioReactContext.Provider value={contextValue}>
-                  <WorkSpaceMenu
-                    workspaceIds={['w1']}
-                    studioResource={resource}
-                    onListUpdate={vi.fn}
-                  ></WorkSpaceMenu>
-                </StudioReactContext.Provider>
-              </QueryClientProvider>
-            </NexusProvider>
-          </ConnectedRouter>
-        </Provider>
-      );
-    });
+    if (component) {
+      component.unmount();
+    }
 
-    await waitFor(async () => {
-      const workSpaceAction = (await screen.getAllByText(
-        'Workspace'
-      )[0]) as HTMLButtonElement;
-      fireEvent.click(workSpaceAction);
-    });
-    await act(async () => {
-      const workSpaceAction = (await screen.getAllByText(
-        'Workspace'
-      )[0]) as HTMLButtonElement;
-      fireEvent.click(workSpaceAction);
-    });
-    const editButton = await screen.findByText('Remove');
-    expect(editButton).toBeVisible();
-    await act(async () => {
-      await fireEvent.click(editButton);
-    });
-    await waitFor(async () => {
-      const editForm = await screen.findAllByText('Delete Workspace');
-      expect(editForm).toHaveLength(1);
-    });
-  });
-  it('it displays workspace edit form on clicking edit workspace button', async () => {
-    const { container } = await render(
+    component = await render(
       <Provider store={store}>
         <ConnectedRouter history={history}>
           <NexusProvider nexusClient={nexus}>
@@ -467,96 +120,138 @@ describe('workSpaceMenu', () => {
       </Provider>
     );
 
-    await waitFor(async () => {
-      const workSpaceAction = (await screen.getAllByText(
-        'Workspace'
-      )[0]) as HTMLButtonElement;
-      fireEvent.click(workSpaceAction);
-    });
-    await act(async () => {
-      const workSpaceAction = (await screen.getAllByText(
-        'Workspace'
-      )[0]) as HTMLButtonElement;
-      fireEvent.click(workSpaceAction);
-    });
+    container = component.container;
+
+    user = userEvent.setup();
+  })
+
+  // reset any request handlers that are declared as a part of our tests
+  // (i.e. for testing one-time error scenarios)
+  afterEach(() => {
+    server.resetHandlers();
+    queryClient.clear();
+  });
+
+  // clean up once the tests are done
+  afterAll(() => {
+    server.close()
+  });
+
+
+  it('renders with a single workspace and dashboard ', async () => {
+    const dashboardText = await screen.findByText(/Example/);
+    expect(dashboardText).toBeVisible();
+    const text = await screen.findAllByText(
+      'https://bluebrain.github.io/nexus/vocabulary/apiMappings'
+    );
+    expect(text.length).toBe(9);
+    expect(container).toMatchSnapshot();
+  });
+
+  it('shows edit buttons for a user with  edit access', async () => {
+    const workspaceButton = await screen.findByText('Workspace');
+    expect(workspaceButton).toBeVisible();
+    const dashboardButton = await screen.findByText('Dashboard');
+    expect(dashboardButton).toBeVisible();
+  });
+
+  it('hides edit buttons for a user with out edit access', async () => {
+    server.use(aclHandler);
+
+    await render(
+      <Provider store={store}>
+        <ConnectedRouter history={history}>
+          <NexusProvider nexusClient={nexus}>
+            <QueryClientProvider client={queryClient}>
+              <StudioReactContext.Provider value={contextValue}>
+                <WorkSpaceMenu
+                  workspaceIds={['w1']}
+                  studioResource={resource}
+                  onListUpdate={vi.fn}
+                ></WorkSpaceMenu>
+              </StudioReactContext.Provider>
+            </QueryClientProvider>
+          </NexusProvider>
+        </ConnectedRouter>
+      </Provider>
+    );
+
+    await screen.findAllByText('No dashboards available');
+    const buttons = await screen.findAllByRole('button');
+    expect(buttons).toHaveLength(2);
+  });
+
+  it('it displays workspace action on clicking  workspace action button', async () => {
+    const button1 = await screen.findByText('Workspace');
+    expect(button1).toBeVisible();
+    const button2 = await screen.findByText('Dashboard');
+    expect(button2).toBeVisible();
+    await user.click(button1);
+
+    expect(await screen.findByText('Edit')).toBeInTheDocument();
+    expect(screen.getByText('Add')).toBeInTheDocument();
+    expect(screen.getByText('Remove')).toBeInTheDocument();
+  });
+
+  it('it displays workspace add form on clicking add workspace button', async () => {
+    await screen.findAllByText('Workspace');
+    const workSpaceAction = (await screen.getAllByText(
+      'Workspace'
+    )[0]) as HTMLButtonElement;
+
+    fireEvent.click(workSpaceAction);
+
+    await screen.findByText('Add');
+
+    const editButton = await screen.findByText('Add');
+    await fireEvent.click(editButton);
+
+    const editForm = await screen.findAllByText('Create a new Workspace');
+    expect(editForm).toHaveLength(1);
+  });
+
+  it('it displays workspace remove confirmation  on clicking remove workspace button', async () => {
+    const workSpaceAction = (await screen.findAllByText(
+      'Workspace'
+    ))[0] as HTMLButtonElement;
+    fireEvent.click(workSpaceAction);
+    const editButton = await screen.findByText('Remove');
+    expect(editButton).toBeVisible();
+    await fireEvent.click(editButton);
+    const editForm = await screen.findAllByText('Delete Workspace');
+    expect(editForm).toHaveLength(1);
+  });
+
+  it('it displays workspace edit form on clicking edit workspace button', async () => {
+    const workSpaceAction = (await screen.findAllByText(
+      'Workspace'
+    ))[0] as HTMLButtonElement;
+    fireEvent.click(workSpaceAction);
+
     const editButton = await screen.findByText('Edit');
     expect(editButton).toBeVisible();
-    await act(async () => {
-      await fireEvent.click(editButton);
-      await waitFor(async () => {
-        const editForm = await screen.findAllByText('Edit');
-        expect(editForm).toHaveLength(1);
-      });
-    });
+    await fireEvent.click(editButton);
+
+    const editForm = await screen.findAllByText('Edit');
+    expect(editForm).toHaveLength(1);
   });
 
   it('it displays dashboard edit/add/remove options on clicking dahsboard edit action', async () => {
-    await act(async () => {
-      const { container } = await render(
-        <Provider store={store}>
-          <ConnectedRouter history={history}>
-            <NexusProvider nexusClient={nexus}>
-              <QueryClientProvider client={queryClient}>
-                <StudioReactContext.Provider value={contextValue}>
-                  <WorkSpaceMenu
-                    workspaceIds={['w1']}
-                    studioResource={resource}
-                    onListUpdate={vi.fn}
-                  ></WorkSpaceMenu>
-                </StudioReactContext.Provider>
-              </QueryClientProvider>
-            </NexusProvider>
-          </ConnectedRouter>
-        </Provider>
-      );
-    });
-
-    await waitFor(async () => {
-      const dashboardAction = await screen.findByText('Dashboard');
-      await fireEvent.click(dashboardAction);
-    });
-
-    await act(async () => {
-      const add = await screen.findAllByText('Add');
-      expect(add).toHaveLength(1);
-    });
+    const dashboardAction = await screen.findByText('Dashboard');
+    await fireEvent.click(dashboardAction);
+    const addButton = await screen.findByText('Add');
+    expect(addButton).toBeVisible();
   });
 
   it('it displays dashboard add form when clicked on add button', async () => {
-    await act(async () => {
-      const { container } = await render(
-        <Provider store={store}>
-          <ConnectedRouter history={history}>
-            <NexusProvider nexusClient={nexus}>
-              <QueryClientProvider client={queryClient}>
-                <StudioReactContext.Provider value={contextValue}>
-                  <WorkSpaceMenu
-                    workspaceIds={['w1']}
-                    studioResource={resource}
-                    onListUpdate={vi.fn}
-                  ></WorkSpaceMenu>
-                </StudioReactContext.Provider>
-              </QueryClientProvider>
-            </NexusProvider>
-          </ConnectedRouter>
-        </Provider>
-      );
-    });
+    const dashboardAction = await screen.findByText('Dashboard');
+    await fireEvent.click(dashboardAction);
 
-    await act(async () => {
-      const dashboardAction = await screen.findByText('Dashboard');
-      await fireEvent.click(dashboardAction);
-    });
+    const addButton = await screen.findByText('Add');
+    expect(addButton).toBeVisible();
+    await fireEvent.click(addButton);
 
-    await act(async () => {
-      const add = await screen.findAllByText('Add');
-      expect(add).toHaveLength(1);
-      await fireEvent.click(add[0]);
-    });
-
-    await waitFor(async () => {
-      const dbForm = await screen.findByText('Create Dashboard');
-      expect(dbForm).toBeVisible();
-    });
+    const dbForm = await screen.findByText('Create Dashboard');
+    expect(dbForm).toBeInTheDocument();
   });
 });
